# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GESTAVICOLE SaaS API - Gestion avicole multi-tenant
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /farms:
    get:
      operationId: listFarms
      tags: [farms]
      summary: List farms in tenant
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of farms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmListResponse"
    post:
      operationId: createFarm
      tags: [farms]
      summary: Create a farm
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFarmRequest"
      responses:
        "201":
          description: Farm created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
  /farms/{farmId}:
    get:
      operationId: getFarm
      tags: [farms]
      summary: Get farm by ID
      security:
        - bearerAuth: []
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Farm details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
    put:
      operationId: updateFarm
      tags: [farms]
      summary: Update farm
      security:
        - bearerAuth: []
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFarmRequest"
      responses:
        "200":
          description: Farm updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
    delete:
      operationId: deleteFarm
      tags: [farms]
      summary: Delete farm
      security:
        - bearerAuth: []
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Farm deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /farms/{farmId}/buildings:
    get:
      operationId: listBuildings
      tags: [buildings]
      summary: List buildings in a farm
      security:
        - bearerAuth: []
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of buildings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildingListResponse"
    post:
      operationId: createBuilding
      tags: [buildings]
      summary: Create a building
      security:
        - bearerAuth: []
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBuildingRequest"
      responses:
        "201":
          description: Building created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Building"
...
  /salaries:
    get:
      operationId: listSalaries
      tags: [hr]
      summary: List salary records
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of salaries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalaryListResponse"
    post:
      operationId: createSalary
      tags: [hr]
      summary: Create a salary record
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSalaryRequest"
      responses:
        "201":
          description: Salary record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Salary"
  /salaries/{salaryId}:
    put:
      operationId: updateSalary
      tags: [hr]
      summary: Update a salary record
      security:
        - bearerAuth: []
      parameters:
        - name: salaryId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSalaryRequest"
      responses:
        "200":
          description: Salary record updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Salary"
    delete:
      operationId: deleteSalary
      tags: [hr]
      summary: Delete a salary record
      security:
        - bearerAuth: []
      parameters:
        - name: salaryId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Salary deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Farm:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        location:
          type: string
        capacity:
          type: integer
        tenantId:
          type: string
        managerId:
          type: string
          nullable: true
        managerName:
          type: string
          nullable: true
        buildingsCount:
          type: integer
        activeBatchesCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - location
        - capacity
        - tenantId
        - buildingsCount
        - activeBatchesCount
        - createdAt
    CreateFarmRequest:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        capacity:
          type: integer
        managerId:
          type: string
      required:
        - name
        - location
        - capacity
    UpdateFarmRequest:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        capacity:
          type: integer
        managerId:
          type: string
    FarmListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Farm"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    Building:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        farmId:
          type: string
        type:
          type: string
        capacity:
          type: integer
        currentOccupancy:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - farmId
        - type
        - capacity
        - currentOccupancy
        - createdAt
    CreateBuildingRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        capacity:
          type: integer
      required:
        - name
        - type
        - capacity
    BuildingListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Building"
        total:
          type: integer
      required:
        - data
        - total
...
    Salary:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        userId:
          type: string
        userName:
          type: string
          nullable: true
        userRole:
          type: string
          nullable: true
        month:
          type: integer
        year:
          type: integer
        baseSalary:
          type: number
          format: float
        bonuses:
          type: number
          format: float
        deductions:
          type: number
          format: float
        netSalary:
          type: number
          format: float
        paymentDate:
          type: string
          nullable: true
        paymentStatus:
          type: string
          enum: [EN_ATTENTE, PAYE, ANNULE]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tenantId
        - userId
        - month
        - year
        - baseSalary
        - bonuses
        - deductions
        - netSalary
        - paymentStatus
        - createdAt
    SalaryListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Salary"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    CreateSalaryRequest:
      type: object
      properties:
        userId:
          type: string
        month:
          type: integer
        year:
          type: integer
        baseSalary:
          type: number
          format: float
        bonuses:
          type: number
          format: float
        deductions:
          type: number
          format: float
        paymentDate:
          type: string
        paymentStatus:
          type: string
          enum: [EN_ATTENTE, PAYE, ANNULE]
        notes:
          type: string
      required:
        - userId
        - month
        - year
        - baseSalary
    UpdateSalaryRequest:
      type: object
      properties:
        baseSalary:
          type: number
          format: float
        bonuses:
          type: number
          format: float
        deductions:
          type: number
          format: float
        paymentDate:
          type: string
        paymentStatus:
          type: string
          enum: [EN_ATTENTE, PAYE, ANNULE]
        notes:
          type: string
...
    DeleteResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...